fix:loginページのlogic,viewわけ
